invert train-sim and kotlin_bindings
train-sim/kotlin_bindings -> train-sim
train-sim -> train-sim/libtsim

so that the directory architectures goes kotlin->rust instead of
kotlin->rust->kotlin and we can extract libtsim to another repo and use
it from python or something without python stuff having kotlin stuff in
the way and vice versa well then
