+++ core/settings.gradle
@@ -14,9 +14,7 @@ include 'osrd-geom'
 include 'osrd-railjson'
 include 'envelope-sim'
 include 'osrd-reporting'
-
 include 'train-sim'
-project(':train-sim').projectDir = file('train-sim/kotlin_bindings')
 
 
 dependencyResolutionManagement {
