--- core/train-sim/cross_language_tests/expected_output.json
@@ -1,4 +1,0 @@
-{
-    "formal": "Hello, Jane Doe!",
-    "informal": "Hi, Jane!"
-}
