--- core/train-sim/cross_language_tests/input.json
@@ -1,4 +1,0 @@
-{
-    "name": "Jane",
-    "surname": "Doe"
-}
