--- core/train-sim/kotlin_bindings/.dockerignore
@@ -1,1 +1,0 @@
-build
