+++ core/train-sim/libtsim/.dockerignore
