+++ core/train-sim/libtsim/.gitignore
