+++ core/train-sim/libtsim/Cargo.toml
@@ -1,7 +1,25 @@
 [package]
-name = "libtsim"
-version = "3.0.0"
-edition = "2024"
+name = "train-sim"
+version = "0.1.0"
+edition = "2021"
 license = "LGPL-3.0"
 
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[workspace]
+members = [".", "libtsim", "uniffi-bindgen"]
+
+[workspace.dependencies]
+libtsim = { version = "3.0.0", path = "libtsim" }
+uniffi = "0.29"
+
 [dependencies]
+libtsim.workspace = true
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+thiserror = "1.0"
+uniffi.workspace = true
+
+[build-dependencies]
+uniffi = { workspace = true, features = ["build"] }
