+++ core/train-sim/libtsim/libtsim/src/range_map.rs
