+++ core/train-sim/src/test/kotlin/sncf/osrd/train_sim/test_train_sim.kt
